Fixed test TODOs
